change: error messages, example nav title
