+++ lib/components/ErrorMessages/Forbidden403/Examples.stories.tsx
@@ -1,0 +1,53 @@
+import React from 'react'
+import type { Meta, StoryObj } from '@storybook/react'
+import { Main } from '../../../layouts/Main/Main'
+import { FooterStandard } from '../../Footer/FooterStandard/FooterStandard'
+import { Nav } from '../../Nav/Nav'
+import { NavAsideData } from '../../../data/NavData'
+import { Forbidden403 } from './Forbidden403'
+
+const meta: Meta = {
+  title: 'Components/Error Messages/403 Forbidden/Examples',
+  parameters: {
+    layout: 'fullscreen',
+  },
+}
+
+export default meta
+type Story = StoryObj
+
+const SinglePara = () => {
+  return (
+    <>
+      <p>
+        Nobis voluptatem dolorum et eum doloremque cupiditate velit. Praesentium architecto a distinctio aut
+        reprehenderit ducimus. Perferendis excepturi delectus nihil voluptatem non. Molestiae quas dolores accusamus in.
+        Praesent quis ligula quis nulla malesuada tempor.
+      </p>
+    </>
+  )
+}
+
+export const LayoutExamples: Story = {
+  render: () => (
+    <>
+      <Nav>
+        <Nav.Primary>
+          <Nav.Logo title="Raven Design System" link="/" />
+          {/* <Nav.Menu menu={NavDataSingle} /> */}
+          <Nav.Aside menu={NavAsideData} />
+        </Nav.Primary>
+      </Nav>
+
+      <Main maxWidth="5xl">
+        <Forbidden403 />
+
+        <h2>Header Two</h2>
+        <SinglePara />
+        <SinglePara />
+      </Main>
+
+      <FooterStandard />
+    </>
+  ),
+}
