+++ lib/components/Testimonial/Examples.stories.TODO.tsx
@@ -33,7 +33,7 @@ export const LayoutExamples: Story = {
     <>
       <Nav>
         <Nav.Primary>
-          <Nav.Logo title="Web Services" link="https://carleton.ca/webservices" />
+          <Nav.Logo title="Raven Design System" link="/" />
           {/* <Nav.Menu menu={NavDataSingle} /> */}
           <Nav.Aside menu={NavAsideData} />
         </Nav.Primary>
