+++ lib/main.ts
@@ -18,7 +18,7 @@ export { FooterBasic } from './components/Footer/FooterBasic/FooterBasic'
 export { FooterCookie } from './components/Footer/FooterCookie/FooterCookie'
 export { FooterLogoLinks } from './components/Footer/FooterLogoLinks/FooterLogoLinks'
 export { FooterStandard } from './components/Footer/FooterStandard/FooterStandard'
-export { Forbidden403 } from './components/Forbidden403/Forbidden403'
+export { Forbidden403 } from './components/ErrorMessages/Forbidden403/Forbidden403'
 export { FullBanner } from './components/FullBanner/FullBanner'
 export { Figure } from './components/Figure/Figure'
 export { Icon } from './components/Icon/Icon'
@@ -33,7 +33,7 @@ export { Meta } from './components/Meta/Meta'
 export { Modal } from './components/Modal/Modal'
 export { MultiDayCalendar } from './components/MultiDayCalendar/MultiDayCalendar'
 export { Nav } from './components/Nav/Nav'
-export { NotFound404 } from './components/NotFound404/NotFound404'
+export { NotFound404 } from './components/ErrorMessages/NotFound404/NotFound404'
 export { PageHeader } from './components/PageHeader/PageHeader'
 export { Pagination } from './components/Pagination/Pagination'
 export { Placeholder } from './components/Placeholder/Placeholder'
